# Conflicts:
#	ADJ-Internship/BusinessService/Implementations/PurchaseOrderService.cs
#	ADJ-Internship/DataAccess/Migrations/ApplicationDbContextModelSnapshot.cs
#	ADJ-Internship/DataModel/OrderTrack/Order.cs
#	ADJ-Internship/DataModel/OrderTrack/OrderDetail.cs
#	ADJ-Internship/Repository/Implementations/OrderDetailRepository.cs
#	ADJ-Internship/Repository/Implementations/OrderRepository.cs
#	ADJ-Internship/Repository/Implementations/ProgressCheckRepository.cs
#	ADJ-Internship/Repository/Interfaces/IOrderDetailRepository.cs
#	ADJ-Internship/Repository/Interfaces/IOrderRepository.cs
#	ADJ-Internship/Repository/Interfaces/IProgressCheckRepository.cs
#	ADJ-Internship/WebApp/Views/Shared/_Layout.cshtml

+++ ADJ-Internship/DataModel/OrderTrack/OrderDetail.cs
@@ -29,7 +29,7 @@ namespace ADJ.DataModel.OrderTrack
         [Required]
         public decimal Quantity { get; set; }
 
-        public float ReviseQuantity { get; set; }
+        public decimal ReviseQuantity { get; set; }
 
         [Required]
         public float Cartons { get; set; }
